+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## Features
-- Adjust text and icon for `Stop After Next Quest` button - Kage
-## Bug fixes
+- Feature: Make It Rain 2026 event -alydev
+- Feature: Adjust text and icon for `Stop After Next Quest` button - Kage
+- Bug: Refactor and code style -alydev